Introducing the generic profile loader
